Hab das mit den Buttons gelöst. Kannst du dir den VideoPlayer anschauen?
