+++ package.json
@@ -21,7 +21,8 @@
     "react": "18.2.0",
     "react-native": "0.72.4",
     "react-native-paper": "^5.10.3",
-    "react-native-progress-bar": "^0.1.2"
+    "react-native-progress-bar": "^0.1.2",
+    "react-native-video": "^5.2.1"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
