+++ screens/Agreement.js
@@ -1,11 +1,8 @@
 import React from 'react';
 import { View, Text, StyleSheet, TouchableOpacity } from 'react-native';
 import {useNavigation} from '@react-navigation/native';
-import NextButton from '../components/NextButton';
-import BackButton from '../components/BackButton';
+import BackNextButton from '../components/BackNextButton';
 import { styles } from './Information';
-
-
 
 export default function Agreement() {
 
@@ -19,8 +16,7 @@ export default function Agreement() {
         <Text>Agreement</Text>
       </TouchableOpacity>
       <View style={styles.footer}>
-        <BackButton label="Back" onPress={() => navigation.goBack()}/>
-        <NextButton label='Next' onPress={() => {navigate('FAQ')}}/>
+        <BackNextButton onBackPress={() => navigation.goBack()} onNextPress={() => {navigate('FAQ')}}/>
       </View>
     </View>
   );
